Add files via upload
After the increase of interest rates by the ECB at the beginning of 2023, it is interesting to check how the different industrial sectors were affected. This analysis leads to the conclusion that the most affected activity was the Real Estate (i.e. it has become more risky and more likely to experience defaults).